refactor: extract Calendar components and hooks
